build script update for linux

+++ src/common-ext.h
@@ -17,8 +17,8 @@ static bool label_fps_indicate(int64_t now,BiTimer* timer)
   char text[1024];
   snprintf(text,1024,"FPS:%.2f CB:%lld RE:%lld ma:%.2f OT:%.2f / %d,%d / %d,%d",
     context->profile.stats.fps,
-    context->profile.stats.time_spent_on_callback,
-    context->profile.stats.time_spent_on_rendering,
+    (long long)context->profile.stats.time_spent_on_callback,
+    (long long)context->profile.stats.time_spent_on_rendering,
     context->profile.stats.matrix_updated,
 
     context->profile.stats.fps == 0 ? 0 : 1000.0 / context->profile.stats.fps - context->profile.stats.time_spent_on_callback - context->profile.stats.time_spent_on_rendering,
